sperate styling into comp folders
