+++ my-app/src/components/About/aboutC.scss
