+++ my-app/src/components/Community/communityC.scss
