+++ my-app/src/components/Hero/Hero.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import { useState, useEffect } from "react";
 import "./hero.scss";
 
 const Hero = () => {
