+++ my-app/src/components/HowItWorks/HowItWorks.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./howItWorks.scss";
 
 const HowItWorks = () => {
