+++ my-app/src/pages/SignIn.tsx
@@ -1,168 +1,10 @@
-import React, { useState } from "react";
-import "./signIn.scss";
+import SignInC from "../components/SignIn/SignInC";
 
 const SignIn = () => {
-  const [formData, setFormData] = useState({
-    email: "",
-    password: "",
-    rememberMe: false,
-  });
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value, type, checked } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: type === "checkbox" ? checked : value,
-    }));
-  };
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    // Handle sign in logic here
-    console.log("Sign in:", formData);
-  };
-
   return (
-    <div className="signin">
-      <div className="signin__container">
-        <div className="signin__content">
-          <div className="signin__header">
-            <h1 className="signin__title">Welcome Back</h1>
-            <p className="signin__subtitle">
-              Sign in to continue your journey toward meaningful connections
-            </p>
-          </div>
-
-          <form className="signin__form" onSubmit={handleSubmit}>
-            <div className="signin__form-group">
-              <label htmlFor="email" className="signin__label">
-                Email Address
-              </label>
-              <input
-                type="email"
-                id="email"
-                name="email"
-                value={formData.email}
-                onChange={handleInputChange}
-                className="signin__input"
-                placeholder="Enter your email"
-                required
-              />
-            </div>
-
-            <div className="signin__form-group">
-              <label htmlFor="password" className="signin__label">
-                Password
-              </label>
-              <input
-                type="password"
-                id="password"
-                name="password"
-                value={formData.password}
-                onChange={handleInputChange}
-                className="signin__input"
-                placeholder="Enter your password"
-                required
-              />
-            </div>
-
-            <div className="signin__form-options">
-              <label className="signin__checkbox-label">
-                <input
-                  type="checkbox"
-                  name="rememberMe"
-                  checked={formData.rememberMe}
-                  onChange={handleInputChange}
-                  className="signin__checkbox"
-                />
-                <span className="signin__checkbox-text">Remember me</span>
-              </label>
-              <a href="#" className="signin__forgot-link">
-                Forgot password?
-              </a>
-            </div>
-
-            <button type="submit" className="signin__submit">
-              Sign In
-            </button>
-          </form>
-
-          <div className="signin__divider">
-            <span className="signin__divider-text">or</span>
-          </div>
-
-          <div className="signin__social">
-            <button className="signin__social-button signin__social-button--google">
-              <svg className="signin__social-icon" viewBox="0 0 24 24">
-                <path
-                  d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-                  fill="#4285F4"
-                />
-                <path
-                  d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-                  fill="#34A853"
-                />
-                <path
-                  d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-                  fill="#FBBC05"
-                />
-                <path
-                  d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-                  fill="#EA4335"
-                />
-              </svg>
-              Continue with Google
-            </button>
-          </div>
-
-          <div className="signin__social">
-            <button className="signin__social-button signin__social-button--outlook">
-              <svg
-                className="signin__social-icon"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-              >
-                <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm0 18c-4.41 0-8-3.59-8-8s3.59-8 8-8 8 3.59 8 8-3.59 8-8 8z" />
-                <path d="M12 6c-3.31 0-6 2.69-6 6s2.69 6 6 6 6-2.69 6-6-2.69-6-6-6zm0 10c-2.21 0-4-1.79-4-4s1.79-4 4-4 4 1.79 4 4-1.79 4-4 4z" />
-              </svg>
-              Continue with Microsoft Outlook
-            </button>
-          </div>
-
-          <div className="signin__social">
-            <button className="signin__social-button signin__social-button--github">
-              <svg
-                className="signin__social-icon"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-              >
-                <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z" />
-              </svg>
-              Continue with GitHub
-            </button>
-          </div>
-
-          <div className="signin__footer">
-            <p className="signin__footer-text">
-              Don't have an account?{" "}
-              <a href="/signup" className="signin__footer-link">
-                Sign up
-              </a>
-            </p>
-          </div>
-        </div>
-
-        <div className="signin__image">
-          <div className="signin__image-content">
-            <h2 className="signin__image-title">Connect for Real</h2>
-            <p className="signin__image-subtitle">
-              Join thousands of people building meaningful connections through
-              shared experiences.
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
+    <>
+      <SignInC />
+    </>
   );
 };
