--- my-app/src/pages/login.tsx
@@ -1,11 +1,0 @@
-
-function Login () {
-    return (
-      <>
-        
-      </>
-    );
-  }
-  
-  export default Login;
-
