--- my-app/src/pages/signIn.scss
@@ -1,314 +1,0 @@
-.signin {
-  min-height: 100vh;
-  background: linear-gradient(135deg, #fefefe 0%, #f8f6f3 100%);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 2rem 0;
-
-  @media (max-width: 768px) {
-    padding: 1rem 0;
-  }
-}
-
-.signin__container {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  max-width: 1200px;
-  width: 100%;
-  background: white;
-  border-radius: 2rem;
-  overflow: hidden;
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1);
-  border: 1px solid rgba(212, 165, 116, 0.15);
-
-  @media (max-width: 1024px) {
-    grid-template-columns: 1fr;
-    max-width: 500px;
-  }
-
-  @media (max-width: 768px) {
-    border-radius: 1rem;
-    margin: 0 1rem;
-  }
-}
-
-.signin__content {
-  padding: 3rem 2.5rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-
-  @media (max-width: 768px) {
-    padding: 2rem 1.5rem;
-  }
-}
-
-.signin__header {
-  text-align: center;
-  margin-bottom: 2.5rem;
-
-  @media (max-width: 768px) {
-    margin-bottom: 2rem;
-  }
-}
-
-.signin__title {
-  font-size: 2.5rem;
-  font-weight: 700;
-  color: #2d3748;
-  margin-bottom: 0.75rem;
-
-  @media (max-width: 768px) {
-    font-size: 2rem;
-  }
-}
-
-.signin__subtitle {
-  font-size: 1rem;
-  color: #4a5568;
-  line-height: 1.5;
-}
-
-.signin__form {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.signin__form-group {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.signin__label {
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: #2d3748;
-}
-
-.signin__input {
-  padding: 1rem 1.25rem;
-  border: 2px solid #e2e8f0;
-  border-radius: 0.75rem;
-  font-size: 1rem;
-  color: #2d3748;
-  background: white;
-  transition: all 0.3s ease;
-
-  &:focus {
-    outline: none;
-    border-color: #d4a574;
-    box-shadow: 0 0 0 3px rgba(212, 165, 116, 0.1);
-  }
-
-  &::placeholder {
-    color: #a0aec0;
-  }
-}
-
-.signin__form-options {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 0.5rem;
-}
-
-.signin__checkbox-label {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  cursor: pointer;
-}
-
-.signin__checkbox {
-  width: 1rem;
-  height: 1rem;
-  accent-color: #d4a574;
-  cursor: pointer;
-}
-
-.signin__checkbox-text {
-  font-size: 0.875rem;
-  color: #4a5568;
-}
-
-.signin__forgot-link {
-  font-size: 0.875rem;
-  color: #d4a574;
-  text-decoration: none;
-  font-weight: 500;
-  transition: color 0.3s ease;
-
-  &:hover {
-    color: #b8860b;
-  }
-}
-
-.signin__submit {
-  background: linear-gradient(135deg, #d4a574, #b8860b);
-  color: white;
-  border: none;
-  padding: 1rem 2rem;
-  border-radius: 0.75rem;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  margin-top: 1rem;
-
-  &:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 8px 20px rgba(212, 165, 116, 0.3);
-  }
-
-  &:active {
-    transform: translateY(0);
-  }
-}
-
-.signin__divider {
-  position: relative;
-  text-align: center;
-  margin: 2rem 0;
-
-  &::before {
-    content: "";
-    position: absolute;
-    top: 50%;
-    left: 0;
-    right: 0;
-    height: 1px;
-    background: #e2e8f0;
-  }
-}
-
-.signin__divider-text {
-  background: white;
-  padding: 0 1rem;
-  color: #a0aec0;
-  font-size: 0.875rem;
-  position: relative;
-  z-index: 1;
-}
-
-.signin__social {
-  margin-bottom: 2rem;
-}
-
-.signin__social-button {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.75rem;
-  padding: 1rem 2rem;
-  border: 2px solid #e2e8f0;
-  border-radius: 0.75rem;
-  background: white;
-  color: #2d3748;
-  font-size: 1rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.3s ease;
-
-  &:hover {
-    border-color: #d4a574;
-    background: rgba(212, 165, 116, 0.05);
-  }
-
-  &--outlook {
-    &:hover {
-      border-color: #0078d4;
-      background: rgba(0, 120, 212, 0.05);
-    }
-  }
-
-  &--github {
-    &:hover {
-      border-color: #333;
-      background: rgba(51, 51, 51, 0.05);
-    }
-  }
-}
-
-.signin__social-icon {
-  width: 1.5rem;
-  height: 1.5rem;
-}
-
-.signin__footer {
-  text-align: center;
-  margin-top: auto;
-}
-
-.signin__footer-text {
-  font-size: 0.875rem;
-  color: #4a5568;
-}
-
-.signin__footer-link {
-  color: #d4a574;
-  text-decoration: none;
-  font-weight: 600;
-  transition: color 0.3s ease;
-
-  &:hover {
-    color: #b8860b;
-  }
-}
-
-.signin__image {
-  background: linear-gradient(135deg, #d4a574, #b8860b);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 3rem 2rem;
-  position: relative;
-  overflow: hidden;
-
-  &::before {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background: url("https://images.unsplash.com/photo-1571019613454-1cb2f99b2d8b?w=800&h=600&fit=crop")
-      center/cover;
-    opacity: 0.1;
-  }
-
-  @media (max-width: 1024px) {
-    display: none;
-  }
-}
-
-.signin__image-content {
-  text-align: center;
-  color: white;
-  position: relative;
-  z-index: 1;
-}
-
-.signin__image-title {
-  font-size: 2.5rem;
-  font-weight: 700;
-  margin-bottom: 1rem;
-
-  @media (max-width: 768px) {
-    font-size: 2rem;
-  }
-}
-
-.signin__image-subtitle {
-  font-size: 1.125rem;
-  line-height: 1.6;
-  opacity: 0.9;
-
-  @media (max-width: 768px) {
-    font-size: 1rem;
-  }
-}
